Update eventmachine to work on Ruby v2.2
I was seeing a bunch of compilation errors when bundler was trying to install
eventmachine. I noticed that the project was originally run on Ruby v2.0.0p195
and a fix [1] was made to eventmachine in v1.0.4 to make it work with
Ruby v2.2.

I decided to update eventmachine to the latest version (v1.0.7) for good
measure.

[1]: https://github.com/eventmachine/eventmachine/blob/master/CHANGELOG.md#104-december-19-2014